py/services/statistics/pyproject.toml
[tool.poetry]
name = "statistics"
version = "0.1.0"
description = ""
authors = ["Aleksey Dalekin <adalekin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aiokafka = "^0.7.2"
cassandra-driver = "^3.25.0"
environs = "9.3.2"
fastapi = "^0.68.1"
httpx = "^0.19.0"
schemas = {path = "../../lib/schemas"}
uvicorn = {extras = ["standard"], version = "^0.15.0"}

[tool.poetry.dev-dependencies]
asgi-lifespan = "^1.0.1"
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
pytest-cov = "^2.12.1"
pytest-env = "^0.6.2"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
